Set up basic webpage structure
An index html file with linked javascript file and css file are set up.
They are tested and are properly connected.
An initial container div is placed in the html, within which the
etch-a-sketch grid will later be populated using Javascript.

Basic style reset is added to the top of the css file.