Create function to populate sketch grid
With this commit a function is written in the script(.js) to create a
grid of divs.
It does this by taking a number as its argument and then looping to
create divs, which it then appends as a child to the .container div
in the DOM.

A css class is also appended to each div in the script function.
This class is currently used, just as a test to see that the divs
actually do appear visibly in the web-page.