Make grid positioning dynamic
In the last commit, the main container div had been set up as a CSS grid
so that it could be styled in a way, whereby each tile in the grid would
be given an equal space.
With this commit, this job is passed to a Javascript function, which
takes a numberOfTiles given to it from its calling function (makeGrid)
and uses that to add the gridTemplate style to the container DOM element
with the correct number of equally repeating columns and rows.

All other CSS grid related styling is still left up to the CSS, as this
can remain static, no matter how many tiles in the grid.