Make grid a centred square
Before this commit, the grid would be sized within the window, just so
that it had 10% padding on all sides.
It would usually be a rectangle, therefore, unless the window was
completely square.

This commit makes the grid a square, no matter what the size of the
window is and it also will remain centred, vertically and horizontally.
This commit also provides the function for this state to be maintained
during resize events.

Centering of the grid is performed by Javascript, which finds the window
height and width and depending on which is smallest, makes the largest
square out of the grid, by applying the relevant styles to the DOM
element.
The function responsible for this calculation is also connected
to a window "resize" event listener.
Padding is adjusted in the CSS to no longer be dynamic, as this messes
with the resize function in a detrimental way.

A centering-device is placed around the grid's div, because this allows
to grid to be centered within it, both horizontally and vertically,
by making the centering-device a flex-container.
HTML, BODY and the centering-device itself, all need to be given a width
and, more importantly, a height of 100%, because otherwise, when the
page is resized below a certain width, the HTML, body and centering
device will not extened to the full height of the page.