Make grid tiles paintable
With this commit, tiles are given a little more styling to make them
appear visible, but blank before painting.

Also with this commit, they are made paintable.
To do this, an event listener is given to the sketchBoard element which
listens for mouseover events and then calls the function paintTiles.
The paintTiles function checks to see if the event.target has a class
of tile (which it should if it is a tile) and, if so, appends a class of
painted to it also.
A CSS ruleset exists for painted tiles, so that once the class is added,
that tile will appear to be painted.