Add grid reset button
This commit adds a grid reset button to the HTML.
This button is connected to a javascript function via a click event
handler.
The function uses a confirmation dialogue to ensure the user does not
accidentally delete their work. If the confirmation is positive, the
painted class is removed from all tiles.

Also add a little additional styling to painted tiles,
to give them a more rewarding feel.