Add Grid Resize button and input
With this commit, a grid resize button and input are added to the page.
This is so that a user can choose to work with different sizes of grid.

The control buttons and input are now put into a separate div to
keep them together.
The positioning of things will need to be adjusted in a future commit,
as there is now a bug that occurs with the sizing of the grid when the
window size is changed.

The original createGrid function is the main function at work in the
resizeGrid function.
The original function has been adjusted only slightly to make this
possible. At the beginning of createGrid, there is now a statement that
erases the innerHTML of the container div.
This is so that on successive invocations of the createGrid function,
more and more elements or not added to the grid, so as to bog it down
with an uneven amount of tiles.

The resizeGrid function involves a check to make sure the size is
between 8 and 64.
This is because a grid smaller than 8 tiles, seems to be pointlessly low
in definition and a grid much larger than 64 tiles does not select tiles
very well with the mouse any longer.