Add Paint Mode Selection buttons and functionality
With this commit, the user is able to select between a solid or gradual
paint mode via buttons on the side.
When gradual is selected, the color will gradually darken until finally
becoming black.
Only one option can be selected at a time and, as a result of styling,
the selected paint mode remains toggled on.

Within the Javascript, a paint-mode switch is defined and adjusted
via the togglePaintMode function, which is connected to event listeners
on both paint mode buttons.

The paintTiles function is completely changed now with an if statement.
A different paint function occurs, depending on which paint mode is
selected.
To assist with the calculations in the new gradual paint-mode, an RGB
alpha-increment function has been added and referred to
in the paintTiles function.

The resetGrid function was also previously just removing the "painted"
class from every tile, but it now also clears any directly applied style
as otherwise, any style applied directly by the gradual side of the
paintTiles function, would not reset when then button is clicked.