Give functionality to precision toggle button
Previously, the board's aesthetic look, created with a combination of
the board's grid-gap and the individual tiles' border-radius, makes it
slightly less likely for the mouseover event to reach them and thus
makes painting slightly less precise.
As such, in the last commit a button was added which would change this
styling for increased accuracy.
With this commit, that button is given full functionality.

New style rules are made for the grid container and tiles when they are
given the class of precise.
Within the script, an event listener is given to the button and a new
function connected to it, for toggling Paint Precision, by adding or
removing the style class, to both the tiles and the grid container.
The state of precision is stored in the boolean variable paintPrecision.

An if statement is also added to the createGrid function, so that if the
grid should be resized with precision still turned on, the tiles will be
appended a class of precise upon creation of the new grid.

The element selector statements are also given semicolons for neatness.