Add resize with enter key functionality
Before this commit, the only way to resize the gird was to input a
number into the resize input and then press the resize button.
With this commit, an event listener is given to the resize input, which
listens for a keypress, then passes the event to a resize handler
function.
This handler function screens the keypress event and if it is the enter
key, it calls resizeGrid.

Also in this commit: the text to which the precision button changes upon
click is adjusted slightly to reflect its nature more appropriately.