Add functionality to reset-confirm button
Previously, when resetting the grid, a confirmation dialogue would
appear, which the user would have to confirm in order to clear the grid.
One issue with this was that, when clicking OK, the user would now be
stuck with their cursor in the centre of the grid, unable to remove
their cursor from the grid without painting a line through the grid.

To solve this, a reset-confirm button is implemented.
The reset-confirm button itself was added in a previous commit, but is
made functional with this commit.
Handling of the resetGrid function itself is now given to this new
reset-confirm button, while the original button is instead connected to
a function which makes the confirm button visible in the document
and simultaneously arms the confirm button with an event listener,
linking it to the original resetGrid functionality.
The resetGrid is tweaked to accommodate this adjustment and lines are
added to the end of it to remove the visibility of the
reset-confirm button and to disconnect its event listener
(i.e disarm it).