Style controls panel
With this commit, a font is added to a new fonts folder and then rules
are added in the CSS that apply this font for usage in the web-page.
The control buttons are styled and their panel styled in such a way as
that they are less noticeable and intrusive while not in use, yet still
accessible when needed.
The responsiveness and feel of the controls are also more aesthetically
pleasing for the user now.

Due to adjustments in flex properties of the panel and its contents, the
controls now also stand up quite well to resizing of the document.